+++ needle/core/framework/module.py
@@ -1,10 +1,7 @@
 from __future__ import print_function
 import os
 import time
-import json
-import plistlib
 import textwrap
-from pprint import pprint
 
 from ..framework.framework import Framework, FrameworkException
 from ..framework.options import Options
@@ -160,29 +157,36 @@ class BaseModule(Framework):
         """Pretty print output coming from command execution. Also save it to file if specified"""
         def print_screen(content):
             content_type = type(content)
-            if content_type is dict or content_type is plistlib._InternalDict: pprint(content, indent=4)
-            elif content_type is list: map(print_screen, content)
-            else: print('\t%s' % content.strip())
+            if content_type is dict:
+                Utils.dict_print(content)
+            elif Utils.is_plist(content):
+                Utils.plist_print(content)
+            elif content_type is list:
+                map(print_screen, content)
+            else:
+                print('\t%s' % content.strip())
 
         def print_file(content):
             content_type = type(content)
-            if content_type is dict or content_type is plistlib._InternalDict:
-                try: json.dump(content, fp)
-                except TypeError: pass
+            if content_type is dict:
+                Utils.dict_write_to_file(content, fp)
+            elif Utils.is_plist(content):
+                Utils.plist_write_to_file(content, fp)
             elif content_type is list:
-                for line in content: fp.write('%s\n' % line)
+                for line in content: fp.write('%s\n' % line.strip())
             else:
-                fp.write('%s\n' % content)
+                fp.write('%s\n' % content.strip())
 
         if txt:
             # Print to screen
-            if not silent: print_screen(txt)
+            if not silent:
+                print_screen(txt)
             # Saving to file
             if outfile:
                 if type(outfile) is not str:
                     self.printer.error("Please specify a valid path if you want to save to file")
                 else:
-                    self.printer.info("Saving output to file: %s" % outfile)
+                    self.printer.info("Saving output to file: {}".format(outfile))
                     with open(outfile, 'w') as fp:
                         print_file(txt)