+++ needle/core/utils/utils.py
@@ -1,5 +1,10 @@
 import os
 import re
+import io
+import json
+import biplist
+import plistlib
+from pprint import pprint
 
 
 # ======================================================================================================================
@@ -70,3 +75,60 @@ class Utils(object):
         control_chars = ''.join(map(unichr, range(0, 32) + range(127, 160)))
         control_char_re = re.compile('[%s]' % re.escape(control_chars))
         return control_char_re.sub('', text)
+
+    # ==================================================================================================================
+    # DATA STRUCTURE UTILS
+    # ==================================================================================================================
+    @staticmethod
+    def merge_dicts(*dict_args):
+        """Given any number of dicts, shallow copy and merge into a new dict."""
+        result = {}
+        for dictionary in dict_args:
+            result.update(dictionary)
+        return result
+
+    @staticmethod
+    def dict_print(text):
+        """Print a dictionary to screen."""
+        pprint(text, indent=4)
+
+    @staticmethod
+    def dict_write_to_file(text, fp):
+        """Print a dictionary to file."""
+        try:
+            json.dump(text, fp, indent=4)
+        except TypeError as e:
+            raise Exception(e)
+
+    # ==================================================================================================================
+    # PLIST UTILS
+    # ==================================================================================================================
+    @staticmethod
+    def is_plist(text):
+        """Checks if text is a plist."""
+        content_type = type(text)
+        return content_type is plistlib._InternalDict
+
+    @staticmethod
+    def plist_print(text):
+        """Print a plist to screen."""
+        Utils.dict_print(text)
+
+    @staticmethod
+    def plist_read_from_file(path):
+        """Read a plist from file."""
+        try:
+            plist = biplist.readPlist(path)
+            return plist
+        except (biplist.InvalidPlistException, biplist.NotBinaryPlistException), e:
+            raise Exception("Failed to parse plist file: {}".format(e))
+
+    @staticmethod
+    def plist_read_from_string(text):
+        """Read a plist from string."""
+        Utils.plist_read_from_file(io.BytesIO(text))
+
+    @staticmethod
+    def plist_write_to_file(text, fp):
+        """Write a plist to file."""
+        Utils.dict_write_to_file(text, fp)