+++ needle/modules/hooking/frida/script_dump-ui.py
@@ -13,14 +13,13 @@ class Module(FridaScript):
 
     JS = '''\
 if(ObjC.available) {
-    ObjC.schedule(ObjC.mainQueue, () => {
+    ObjC.schedule(ObjC.mainQueue, function() {
         const window = ObjC.classes.UIWindow.keyWindow();
         const ui = window.recursiveDescription().toString();
         send(ui);
     });
 } else {
     console.log("Objective-C Runtime is not available!");
-}
     '''
 
     # ==================================================================================================================