+++ CHANGELOG.md
@@ -6,7 +6,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 #### Added
+- **[CORE]** Improved metadata parsing for extensions
+- **[MODULE]** Frida Script: Anti Hooking Checks (`hooking/frida/script_anti-hooking-check`) _[from @HenryHoggard]_
+- **[MODULE]** Checksums (`binary/checksums`) _[from @HenryHoggard]_
 #### Fixed
+- **[CORE]** App metadata: creation of binary path from MobileInstallation.plist
+- **[CORE]** Plist wrapper using biplist
+- **[MODULE]** Frida Script: print view hierarchy (`hooking/frida/script_dump-ui`) _[from @HenryHoggard]_
+- **[MODULE]** Improved SQLite DB identification by reducing false positives and false negatives _[from @HenryHoggard]_
 #### Removed